refactor: refactored the API into related handler files
